Queue: half-star ratings + rich community seed data
Half-star ratings:
- Movie.rating changed from Integer to Float
- New _macros.html: stars() macro with CSS half-star gradient rendering, rating_select() macro with 0.5-5.0 options
- Updated all 8 star-display locations (user_detail, film_detail, movie_detail, feed, diary, index) to use macro
- Edit and add-movie forms now show 10 half-star options

Community seed:
- 126 authentic reviews across 22 users, each matching their film persona
- 22 additional follow relationships for a denser social graph
- 622 review likes between followers (75% like rate)

+++ models.py
@@ -128,7 +128,7 @@ class Movie(db.Model):
     title      = db.Column(db.String(120), nullable=False)
     user_id    = db.Column(db.Integer, db.ForeignKey("user.id"), nullable=False)
     film_id    = db.Column(db.Integer, db.ForeignKey("film.id"), nullable=True)
-    rating     = db.Column(db.Integer, nullable=True)
+    rating     = db.Column(db.Float, nullable=True)
     year       = db.Column(db.String(10), nullable=True)
     director   = db.Column(db.String(120), nullable=True)
     plot       = db.Column(db.Text, nullable=True)
